fix: sales module - 10 bugs fixed + activate sidebar items
CRITICAL fixes:
- delivery/[id]: guide.order → guide.sales_order (customer name/order# now display)
- invoices API: discount field name mismatch fixed (discount → discount_percent)
- invoices API: payment_method now stored in DB

HIGH fixes:
- delivery-guides API: driver_name + shipping_address now stored
- delivery-guides detail API: added customer join via sales_order
- delivery-guides PUT: now updates driver_name + shipping_address
- delivery/new: shipping_date now sent to API
- api-client: createDeliveryGuide type updated (shipping_date, optional fields)
- customer-returns API: item condition + reason now stored
- customer-returns PUT: added 'cancelled' to valid statuses

MEDIUM fixes:
- sales/page.tsx: delivery guides + invoices show order number instead of UUID
- quotations/new: parseInt → parseFloat for unit price and discount
- invoices/new: parseInt → parseFloat for unit price and discount
- invoices/new: discount field renamed to discount_percent to match API

Sidebar:
- Activated Cotizaciones, Guias de Despacho, Facturas, Devoluciones (removed comingSoon)

+++ apps/web/src/app/dashboard/sales/quotations/new/page.tsx
@@ -216,7 +216,7 @@ export default function NewQuotationPage() {
                             type="number"
                             min="0"
                             value={item.unitPrice}
-                            onChange={(e) => handleItemChange(index, 'unitPrice', parseInt(e.target.value) || 0)}
+                            onChange={(e) => handleItemChange(index, 'unitPrice', parseFloat(e.target.value) || 0)}
                             className="w-full bg-slate-50 border border-slate-200 rounded-lg px-3 py-2 text-sm text-slate-900 text-right focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent transition-colors"
                           />
                         </td>
@@ -226,7 +226,7 @@ export default function NewQuotationPage() {
                             min="0"
                             max="100"
                             value={item.discount}
-                            onChange={(e) => handleItemChange(index, 'discount', parseInt(e.target.value) || 0)}
+                            onChange={(e) => handleItemChange(index, 'discount', parseFloat(e.target.value) || 0)}
                             className="w-full bg-slate-50 border border-slate-200 rounded-lg px-3 py-2 text-sm text-slate-900 text-center focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent transition-colors"
                           />
                         </td>
